+++ ross/rand-clcg4.h
@@ -15,13 +15,13 @@ struct tw_rng
 	/*
 	 * a[j]^{2^w} et a[j]^{2^{v+w}}.
 	 */
-	long	m[4];
-	long	a[4];
-	long	aw[4];
-	long	avw[4];
+	int32_t	m[4];
+	int32_t	a[4];
+	int32_t	aw[4];
+	int32_t	avw[4];
 
 	// the seed..
-	long	seed[4];
+	int32_t	seed[4];
 };
 
 enum SeedType
@@ -31,9 +31,9 @@ enum SeedType
 
 struct tw_rng_stream
 {
-	long	 Ig[4];
-	long	 Lg[4];
-	long	 Cg[4];
+	int32_t	 Ig[4];
+	int32_t	 Lg[4];
+	int32_t	 Cg[4];
 
 	//tw_rng	*rng;
 
@@ -47,8 +47,8 @@ struct tw_rng_stream
 extern tw_rng	*rng_init(int v, int w);
 extern void     rng_set_initial_seed();
 extern void     rng_init_generator(tw_rng_stream * g, SeedType Where);
-extern void     rng_set_seed(tw_rng_stream * g, long * s);
-extern void     rng_get_state(tw_rng_stream * g, long * s);
+extern void     rng_set_seed(tw_rng_stream * g, uint32_t * s);
+extern void     rng_get_state(tw_rng_stream * g, uint32_t * s);
 extern void     rng_write_state(tw_rng_stream * g, FILE *f);
 extern double   rng_gen_val(tw_rng_stream * g);
 extern double   rng_gen_reverse_val(tw_rng_stream * g);
